// src/store/main/system/type.ts
export interface ISystemState {
  usersList: any[]
  usersCounter: number
  roleList: any[]
  roleCounter: number
  goodsList: any[]
  goodsCounter: number
  menuList: any[]
  menuCounter: number
  departmentList: any[]
  departmentCounter: number
  categoryList: any[]
  categoryCounter: number
  formValue: any
  pageInfo: any
}